scripts/requirements.txt
cloudscraper>=1.2.71
selenium>=4.16.0
undetected-chromedriver>=3.5.4
beautifulsoup4>=4.12.2
lxml>=4.9.3
pyyaml>=6.0.1
tqdm>=4.66.1
requests>=2.31.0
fake-useragent>=1.4.0
